Implemented check in from frontend and refactored so details panel is re-rendered
