+++ frontend/src/components/ShiftCard/index.js
@@ -1,118 +1,21 @@
-import dayjs from 'dayjs';
-import isBetween from 'dayjs/plugin/isBetween';
-import CheckInIcon from '../../assets/check-in-icon.png'
-import Plus from '../../assets/plus.png'
-import RequestCoverageIcon from '../../assets/request-coverage.png'
+
 import './index.css'; 
-import { requestToCoverShift } from '../../api/shiftService';
 import { SHIFT_TYPES, COVERAGE_STATUSES } from '../../data/constants';
 
-function ShiftCard({ shift, shiftType, onUpdate, onShiftSelect }) {
-    const currentDate = dayjs();
-    const shiftDay = dayjs(shift.shift_date).format('YYYY-MM-DD');
-    const shiftStart = dayjs(`${shiftDay} ${shift.start_time}`);
-    const shiftEnd = dayjs(`${shiftDay} ${shift.end_time}`);
+function ShiftCard({ shift, shiftType, onShiftSelect, buttonConfig }) {
 
-    const pastShift = currentDate.isAfter(shiftEnd);
-    const currentShift = currentDate.isBetween(shiftStart, shiftEnd, 'minute', '[]');
-
-    const volunteerID = localStorage.getItem('volunteerID');
-
-    const handleCoverShiftClick = async () => {
-        try {
-            const body = {
-                request_id: shift.request_id,
-                volunteer_id: volunteerID,
-            };
-            await requestToCoverShift(body);
-            // notify parent
-            onUpdate();
-        } catch (error) {
-            console.error('Error generating request to cover shift:', error);
-        }
+    const handleShiftSelection = () => {
+        onShiftSelect(shift);
     };
 
-    // TODO Check-in handler for 'my-shifts'
-    const handleCheckInClick = async () => {
-        if (!shift.checked_in && currentShift) {
-            // Perform check-in logic here
-            console.log(`Checking in for shift ${shift.shift_id}`);
-            // Set the state or make API call here to mark the shift as checked in
-        }
-    };
-
-    // TODO
-    const handleRequestCoverageClick = () => {
-        console.log(`Requesting coverage for shift ${shift.shift_id}`);
-        // Add logic for requesting coverage
-    };
-
-    const buttonConfig = {
-        [SHIFT_TYPES.MY_SHIFTS]: {
-            lineColor: 'var(--green)',
-            label: shift.checked_in 
-                ? 'Checked In' 
-                    : currentShift 
-                    ? 'Check In' 
-                        : pastShift
-                        ? 'Missed Shift'
-                            : 'Upcoming',
-            icon: shift.checked_in ? null : currentShift ? CheckInIcon : null,
-            disabled: shift.checked_in || !currentShift,
-            buttonClass: shift.checked_in ? 'checked-in' : '',
-            onClick: handleCheckInClick,
-        },
-        [SHIFT_TYPES.COVERAGE]: {
-            lineColor: 'var(--red)',
-            label: shift.coverage_status === COVERAGE_STATUSES.RESOLVED
-                ? 'Resolved'
-                : shift.coverage_status === COVERAGE_STATUSES.PENDING
-                ? 'Pending Approval'
-                : 'Cover',
-            icon: shift.coverage_status === COVERAGE_STATUSES.OPEN ? Plus : null,
-            disabled: shift.coverage_status === COVERAGE_STATUSES.RESOLVED || shift.coverage_status === COVERAGE_STATUSES.PENDING,
-            onClick: handleCoverShiftClick,
-        },
-        [SHIFT_TYPES.MY_COVERAGE_REQUESTS]: {
-            lineColor: 'var(--yellow)',
-            label: 'Requested Coverage',
-            icon: null,
-            disabled: true,
-            onClick: () => {}, // No action for this state
-        },
-        [SHIFT_TYPES.DEFAULT]: {
+    const { lineColor, label, icon, disabled, buttonClass, onClick } =
+        buttonConfig?.[shiftType] || buttonConfig?.[SHIFT_TYPES.DEFAULT] || {
             lineColor: 'var(--grey)',
             label: 'View Details',
             icon: null,
             disabled: false,
-        },
-        REQUEST_COVERAGE: {
-            lineColor: 'var(--yellow)',
-            label: 'Request Coverage',
-            icon: RequestCoverageIcon,
-            disabled: false,
-            onClick: handleRequestCoverageClick,
-        },
-    };
-
-    const generateButtonsForDetailsPanel = () => {
-        const buttons = [];
-        const primaryButton = buttonConfig[shiftType] || buttonConfig[SHIFT_TYPES.DEFAULT];
-
-        buttons.push(primaryButton);
-        if (shiftType === SHIFT_TYPES.MY_SHIFTS && !shift.checked_in && !pastShift) {
-            buttons.push(buttonConfig.REQUEST_COVERAGE);
-        }
-        return buttons;
-    };
-
-    const handleShiftSelection = () => {
-        const buttons = generateButtonsForDetailsPanel();
-        onShiftSelect({ ...shift, buttons });
-    };
-
-    const { lineColor, label, icon, disabled, buttonClass, onClick } =
-        buttonConfig[shiftType] || buttonConfig[SHIFT_TYPES.DEFAULT];
+            onClick: () => {},
+        }; 
 
     return (
         <div className="shift-card" onClick={handleShiftSelection}>
@@ -133,7 +36,7 @@ function ShiftCard({ shift, shiftType, onUpdate, onShiftSelect }) {
                         <button
                             className={`check-in-button ${buttonClass}`}
                             disabled={disabled}
-                            onClick={onClick}
+                            onClick={() => onClick(shift)}
                         >
                             {icon && <img src={icon} alt="Button Icon" className="card-button-icon" />}
                             {label}
